requirements.txt
# =============================================================================
# MCP AGENT FRAMEWORK (LastMile AI)
# =============================================================================
# Core MCP framework and dependencies
mcp>=1.13.1
mcp-agent>=0.1.27

# Model providers for MCP Agent
mcp-agent[anthropic]>=0.1.27
mcp-agent[openai]>=0.1.27

# =============================================================================
# GOOGLE DRIVE INTEGRATION
# =============================================================================
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
python-docx>=0.8.11

# =============================================================================
# TWITTER/X INTEGRATION
# =============================================================================
tweepy>=4.14.0

# =============================================================================
# LLM PROVIDERS (Direct API access - fallback if MCP Agent not used)
# =============================================================================
openai>=1.3.0
anthropic>=0.48.0

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0

